renamed ethercats strategy
