+++ src/strategies/ethercats-founder-series/index.ts
